+++ axum-core/src/macros.rs
@@ -20,12 +20,10 @@ macro_rules! define_rejection {
 
         impl crate::response::IntoResponse for $name {
             fn into_response(self) -> $crate::response::Response {
-                let body = http_body::Full::from(format!(concat!($body, ": {}"), self.0));
-                let body = $crate::body::boxed(body);
-                let mut res =
-                    http::Response::new(body);
-                *res.status_mut() = http::StatusCode::$status;
-                res
+                (
+                    http::StatusCode::$status,
+                    format!(concat!($body, ": {}"), self.0)
+                ).into_response()
             }
         }
 
@@ -56,7 +54,7 @@ macro_rules! composite_rejection {
         #[non_exhaustive]
         pub enum $name {
             $(
-                #[allow(missing_docs, deprecated)]
+                #[allow(missing_docs)]
                 $variant($variant)
             ),+
         }
@@ -72,7 +70,6 @@ macro_rules! composite_rejection {
         }
 
         $(
-            #[allow(deprecated)]
             impl From<$variant> for $name {
                 fn from(inner: $variant) -> Self {
                     Self::$variant(inner)