+++ axum/src/extension.rs
@@ -1,0 +1,131 @@
+use crate::extract::{rejection::*, FromRequest, RequestParts};
+use crate::response::IntoResponseParts;
+use async_trait::async_trait;
+use axum_core::response::{IntoResponse, Response, ResponseParts};
+use std::ops::Deref;
+
+/// Extractor and response for extensions.
+///
+/// # As extractor
+///
+/// This is commonly used to share state across handlers.
+///
+/// ```rust,no_run
+/// use axum::{
+///     Router,
+///     Extension,
+///     routing::get,
+/// };
+/// use std::sync::Arc;
+///
+/// // Some shared state used throughout our application
+/// struct State {
+///     // ...
+/// }
+///
+/// async fn handler(state: Extension<Arc<State>>) {
+///     // ...
+/// }
+///
+/// let state = Arc::new(State { /* ... */ });
+///
+/// let app = Router::new().route("/", get(handler))
+///     // Add middleware that inserts the state into all incoming request's
+///     // extensions.
+///     .layer(Extension(state));
+/// # async {
+/// # axum::Server::bind(&"".parse().unwrap()).serve(app.into_make_service()).await.unwrap();
+/// # };
+/// ```
+///
+/// If the extension is missing it will reject the request with a `500 Internal
+/// Server Error` response.
+///
+/// # As response
+///
+/// Response extensions can be used to share state with middleware.
+///
+/// ```rust
+/// use axum::{
+///     Extension,
+///     response::IntoResponse,
+/// };
+///
+/// async fn handler() -> impl IntoResponse {
+///     (
+///         Extension(Foo("foo")),
+///         "Hello, World!"
+///     )
+/// }
+///
+/// #[derive(Clone)]
+/// struct Foo(&'static str);
+/// ```
+#[derive(Debug, Clone, Copy)]
+pub struct Extension<T>(pub T);
+
+#[async_trait]
+impl<T, B> FromRequest<B> for Extension<T>
+where
+    T: Clone + Send + Sync + 'static,
+    B: Send,
+{
+    type Rejection = ExtensionRejection;
+
+    async fn from_request(req: &mut RequestParts<B>) -> Result<Self, Self::Rejection> {
+        let value = req
+            .extensions()
+            .get::<T>()
+            .ok_or_else(|| {
+                MissingExtension::from_err(format!(
+                    "Extension of type `{}` was not found. Perhaps you forgot to add it? See `axum::Extension`.",
+                    std::any::type_name::<T>()
+                ))
+            })
+            .map(|x| x.clone())?;
+
+        Ok(Extension(value))
+    }
+}
+
+impl<T> Deref for Extension<T> {
+    type Target = T;
+
+    fn deref(&self) -> &Self::Target {
+        &self.0
+    }
+}
+
+impl<T> IntoResponseParts for Extension<T>
+where
+    T: Send + Sync + 'static,
+{
+    fn into_response_parts(self, res: &mut ResponseParts) {
+        res.insert_extension(self.0);
+    }
+}
+
+impl<T> IntoResponse for Extension<T>
+where
+    T: Send + Sync + 'static,
+{
+    fn into_response(self) -> Response {
+        let mut res = ().into_response();
+        res.extensions_mut().insert(self.0);
+        res
+    }
+}
+
+impl<S, T> tower_layer::Layer<S> for Extension<T>
+where
+    T: Clone + Send + Sync + 'static,
+{
+    type Service = crate::AddExtension<S, T>;
+
+    fn layer(&self, inner: S) -> Self::Service {
+        crate::AddExtension {
+            inner,
+            value: self.0.clone(),
+        }
+    }
+}