+++ axum/src/macros.rs
@@ -57,12 +57,9 @@ macro_rules! define_rejection {
         #[non_exhaustive]
         pub struct $name;
 
-        #[allow(deprecated)]
         impl $crate::response::IntoResponse for $name {
             fn into_response(self) -> $crate::response::Response {
-                let mut res = http::Response::new($crate::body::boxed(crate::body::Full::from($body)));
-                *res.status_mut() = http::StatusCode::$status;
-                res
+                (http::StatusCode::$status, $body).into_response()
             }
         }
 
@@ -100,12 +97,12 @@ macro_rules! define_rejection {
             }
         }
 
-        impl IntoResponse for $name {
+        impl crate::response::IntoResponse for $name {
             fn into_response(self) -> $crate::response::Response {
-                let mut res =
-                    http::Response::new($crate::body::boxed(crate::body::Full::from(format!(concat!($body, ": {}"), self.0))));
-                *res.status_mut() = http::StatusCode::$status;
-                res
+                (
+                    http::StatusCode::$status,
+                    format!(concat!($body, ": {}"), self.0),
+                ).into_response()
             }
         }
 
@@ -136,7 +133,7 @@ macro_rules! composite_rejection {
         #[non_exhaustive]
         pub enum $name {
             $(
-                #[allow(missing_docs, deprecated)]
+                #[allow(missing_docs)]
                 $variant($variant)
             ),+
         }
@@ -152,7 +149,6 @@ macro_rules! composite_rejection {
         }
 
         $(
-            #[allow(deprecated)]
             impl From<$variant> for $name {
                 fn from(inner: $variant) -> Self {
                     Self::$variant(inner)