+++ axum/CHANGELOG.md
@@ -10,6 +10,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - **added:** `middleware::from_fn` for creating middleware from async functions.
   This previously lived in axum-extra but has been moved to axum ([#719])
 - **added:** Document sharing state between handler and middleware (#783])
+- **added:** `Extension<_>` can now be used in tuples for building responses, and will set an
+  extension on the response ([#797])
+- **added:** Implement `tower::Layer` for `Extension` ([#801])
 - **breaking:** `sse::Event` now accepts types implementing `AsRef<str>` instead of `Into<String>`
   as field values.
 - **breaking:** `sse::Event` now panics if a setter method is called twice instead of silently
@@ -52,6 +55,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
         - `PathRejection`
         - `MatchedPathRejection`
         - `WebSocketUpgradeRejection`
+- **breaking:** `Redirect::found` has been removed ([#800])
 - **fixed:** Set `Allow` header when responding with `405 Method Not Allowed` ([#733])
 - **fixed:** Correctly set the `Content-Length` header for response to `HEAD`
   requests ([#734])
@@ -69,6 +73,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [#755]: https://github.com/tokio-rs/axum/pull/755
 [#783]: https://github.com/tokio-rs/axum/pull/783
 [#791]: https://github.com/tokio-rs/axum/pull/791
+[#797]: https://github.com/tokio-rs/axum/pull/797
+[#800]: https://github.com/tokio-rs/axum/pull/800
+[#801]: https://github.com/tokio-rs/axum/pull/801
 
 # 0.4.4 (13. January, 2022)